+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchScreen.kt
@@ -51,6 +51,8 @@ import org.simple.clinic.navigation.v2.Router
 import org.simple.clinic.navigation.v2.ScreenKey
 import org.simple.clinic.navigation.v2.ScreenResultBus
 import org.simple.clinic.navigation.v2.fragments.BaseScreen
+import org.simple.clinic.overdue.download.formatdialog.Download
+import org.simple.clinic.overdue.download.formatdialog.SelectOverdueDownloadFormatDialog
 import org.simple.clinic.summary.OpenIntention
 import org.simple.clinic.summary.PatientSummaryScreenKey
 import org.simple.clinic.util.RuntimeNetworkStatus
@@ -341,6 +343,10 @@ class OverdueSearchScreen : BaseScreen<
 
   override fun setOverdueSearchQuery(searchQuery: String) {
     overdueSearchQueryEditText.setTextAndCursor(searchQuery)
+  }
+
+  override fun openSelectDownloadFormatDialog(appointmentIds: Set<UUID>) {
+    router.push(SelectOverdueDownloadFormatDialog.Key(Download(appointmentIds)))
   }
 
   private fun overdueSearchQueryTextChanges(): Observable<UiEvent> {