+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchViewEffectHandler.kt
@@ -12,6 +12,7 @@ class OverdueSearchViewEffectHandler(
       is OpenPatientSummary -> uiActions.openPatientSummaryScreen(viewEffect.patientUuid)
       is OpenContactPatientSheet -> uiActions.openContactPatientSheet(viewEffect.patientUuid)
       is SetOverdueSearchQuery -> uiActions.setOverdueSearchQuery(viewEffect.searchQuery)
+      is OpenSelectDownloadFormatDialog -> uiActions.openSelectDownloadFormatDialog(viewEffect.selectedAppointmentIds)
     }.exhaustive()
   }
 }